[chalk] Invert Canvas Submission Model
Prior to this changeset, anything that needed to be rendered would
produce canvases and submit them asynchronously, and the pipeline would
choose to either render them with FIFO or Last-Write-Wins semantics.

While this is fantastic for static content and makes it trivial to
provide an "incremental rendering" mode where we can tell the canvas graph is
unmodified and thus skip re-rendering, it makes it very hard to animate
anything that requires continuous motion.

The reasons being that on the FIFO mode, all producers have to be
synchronized to submit canvases exactly as fast as the rendering
pipeline can consume them. This leads to the obvious issue of the
rendering queu building up, and the rendered content lagging behind the
actual system.

On the LWW mode, some canvases will be dropped. In fact, for two subsequent
canvases that are submitted between renders, only the last one will be
rendered. The first submitted canvas will be dropped.

So the choice here is between efficient rendering that is optimized for
screens that do not change often, and flexible rendering that guarantees
all drawn canvases are displayed, and thus allows for complex
animations.

I've decided to pursue the flexible mode, given that an incremental
renderer can be built on top of it.