[chalk] Rewrite with a stronger sense of Rust :crab:
This splits up the codebase into several pieces, and makes better use of
structs to encapsulate the different parts of the system: Compositor,
BeamWriter, BeamReader, CommandIn, CommandOut.

It also helped me really grasp the borrow checker and lifetimes.

No performance improvements were measured until I swapped the
communication to the BEAM from a SegQueue (dynamically allocated thread
safe queue) to just a mspc::channel that will block the reader thread
(which happens to be the one writing to the BEAM!).

All events now include a timestamp for when they occurred, and a
timestamp at which they were sent as well.