[chalk] Rewrite port message handling to handle timestamps
In the process, I've logged, recorded, plotted, and studied the
latencies between:
* Time event happened within the Rust port
* Time event was delivered to stdout
* Time event was received as an port message in Erlang
* Time event was successfuly decoded into Erlang terms
* Time event was dispatched to the chalk_event_server

And started playing around with low-hanging fruit for optimization:

* Setting the process priority to max means the chalk_port will run
  __first__ in the scheduler, no matter what. This guarantees that we
  process incoming port events as soon as they are available.

* Setting port parallelism to false prevents the port process from being
  moved around schedulers, which means we have one less thing to worry
  about that impacts our processing latency (process stealing).

* Binding schedulers to physical CPUs, which lets the OS optimize for
  locality since the same scheduler thread with live in the same CPU
  core.

Something else to try on top of this is to retrieve the topology, and
send to the port the exact CPU core that the port process' scheduler's
thread is running on, so that the port can bind its writer thread to it
too. My hypothesis is that this will guarantee much better locality, and
messages to the port will be delivered considerably faster.