[chalk] Split registration: event and node based
Before this, registration happened on all events, all of the time. After
this, registration for events can happen on an event kind basis, or on a
node-overlap.

The first one is useful when only a subset of all events are relevant:
e.g, if only key-presses are useful, you can register to listen to only
those. If you do want all events, you can register without specifyin
which event kind you are interested in.

The second kind allows you to register events that are spatially
relevant to a particular node in the `chalk_node_tree`. It's a common
yet specific case, where events like `cursor_moved`, can be relevant
only if the coordinates of the mouse are overlapping the coordinates of
your element node.

In any case, if the callback to the event crashes, it will be removed.