[chalk] Uses Quadtree for parallelizing renders
This refactor swaps the original linear render table with a partitioned
table that follows a Quadtree mapping.

What this wins is the ability to parallelize across the Z, Y, and X axis
in __quads__, where each partition of the table can be folded into a
single picture, concurrently, and if possible, parallelly.